Refactor agent architecture: move to `core/agents` with registry
- Migrated all agent modules from `packages/proximal/agents/` to `packages/core/agents/`.
- Introduced an agent registry (`registry.py`) with a `register_agent` decorator.
- Updated CLI command `assist` and orchestrator to import agents from the new path.
- Removed the deprecated `proximal` package structure to streamline the codebase.
- Consolidated planner logic into a class-based `PlannerAgent` with backward-compatible async wrappers.
- Updated imports in providers, tests, and supporting modules to match the new structure.
