+++ packages/core/agents/__init__.py
@@ -1,0 +1,23 @@
+from __future__ import annotations
+from .registry import AGENT_REGISTRY, register_agent
+
+# ensure built-in agents are registered on import
+from .planner import PlannerAgent
+from .chronos import ChronosAgent
+from .guardian import GuardianAgent
+from .mentor import MentorAgent
+from .liaison import LiaisonAgent
+from .scribe import ScribeAgent
+from .focusbuddy import FocusBuddyAgent
+
+from .planner import (
+    PlannerAgent,
+    _json,
+    clarify_llm,
+    integrate_clarifications_llm,
+    plan_llm,
+    prioritize_llm,
+    estimate_llm,
+    package_llm,
+    breakdown_task_llm,
+)
