+++ packages/core/agents/mentor.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
-from . import PlannerAgent, register_agent
+from . import PlannerAgent
+from .registry import register_agent
 
 
 @register_agent("mentor")
