+++ packages/core/integrations/email.py
