+++ packages/core/providers/ollama_provider.py
@@ -1,24 +1,28 @@
+from __future__ import annotations
+
 import httpx
-from typing import Any
 from ..settings import get_settings
+
+from .base import BaseProvider
+from . import register_provider
 
 _SETTINGS = get_settings()
 
 
-class OllamaChat:
-    _model = _SETTINGS.ollama_model
+@register_provider("ollama")
+class OllamaProvider(BaseProvider):
+    """Ollama chat completion provider."""
 
-    @classmethod
-    def _get_url(cls) -> str:
+    def _get_url(self) -> str:
         base = _SETTINGS.ollama_base_url
         if not base.startswith(("http://", "https://")):
             base = f"http://{base}"
         return f"{base}/v1/chat/completions"
 
-    @classmethod
-    async def acomplete(cls, messages: list[dict], tools: Any = None) -> str:
-        url = cls._get_url()
-        payload = {"model": cls._model, "messages": messages}
+    async def chat_complete(self, messages: list[dict], **kwargs: object) -> str:
+        url = self._get_url()
+        payload = {"model": _SETTINGS.ollama_model, "messages": messages}
+        tools = kwargs.get("tools")
         if tools:
             payload["tools"] = tools
         async with httpx.AsyncClient(timeout=120.0) as client:
