+++ packages/core/providers/router.py
@@ -1,30 +1,23 @@
 from __future__ import annotations
-from typing import Any
-from importlib import metadata
+from typing import List, Dict, Any
 from ..settings import get_settings
-from packages.proximal.providers import (
-    PROVIDER_REGISTRY,
-    BaseProvider,
-    register_provider,
-    _load_plugins,
-)
+from . import PROVIDER_REGISTRY
 
-_provider_instance: BaseProvider | None = None
+_settings = get_settings()
+_provider_cache = {}
 
 
-def _get_provider() -> BaseProvider:
-    _load_plugins()
-    global _provider_instance
-    if _provider_instance is None:
-        settings = get_settings()
-        name = settings.provider_name.lower()
-        cls = PROVIDER_REGISTRY.get(name)
-        if not cls:
-            raise ValueError(f"Unknown provider {name}")
-        _provider_instance = cls()
-    return _provider_instance
-
-
-async def chat(messages: list[dict], tools: Any | None = None) -> str:
-    provider = _get_provider()
-    return await provider.chat_complete(messages, tools=tools)
+async def chat(messages: List[Dict[str, Any]], **kwargs: Any) -> str:
+    """Route chat completion to the configured provider."""
+    provider_name = _settings.provider_name.lower()
+    
+    if provider_name not in PROVIDER_REGISTRY:
+        raise ValueError(f"Provider '{provider_name}' not found in registry")
+    
+    # Cache provider instances
+    if provider_name not in _provider_cache:
+        provider_cls = PROVIDER_REGISTRY[provider_name]
+        _provider_cache[provider_name] = provider_cls()
+    
+    provider = _provider_cache[provider_name]
+    return await provider.chat_complete(messages, **kwargs)
